Move infrastructure kuick code to modules
