+++ kuick-api-rest/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    jvmApi project(":kuick-http")
+    jvmApi project(":kuick-api")
+
+    jvmTestImplementation "io.ktor:ktor-server-test-host:$ktor_version"
+    jvmTestImplementation project(":kuick-repositories-squash")
+}
