+++ samples/todo-app2/src/jvmMain/kotlin/kuick/samples/todo2/server.kt
@@ -3,13 +3,17 @@ package kuick.samples.todo2
 import io.ktor.application.Application
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
+import kuick.api.rest.get
+import kuick.api.rest.post
+import kuick.api.rest.restRouting
+import kuick.api.rpc.rpcRouting
 import kuick.client.db.DbClientPool
 import kuick.client.jdbc.JdbcDriver
 import kuick.di.Guice
 import kuick.di.bindPerCoroutineJob
 import kuick.ktor.installContextPerRequest
 import kuick.ktor.installHttpExceptionsSupport
-import kuick.samples.todo2.infrastructure.*
+import kuick.ktor.kuickRouting
 
 suspend fun main(args: Array<String>) {
     embeddedServer(Netty, port = 8080) { module() }.start(wait = true)
@@ -21,7 +25,6 @@ fun Application.module() {
         configure()
     }
 
-    // What's this? Database config also in Guice module
     installContextPerRequest(injector, DbClientPool { JdbcDriver.connectMemoryH2() }) {
         injector.getInstance(TodoRepository::class.java).init()
         injector.getInstance(UserRepository::class.java).init()
