+++ settings.gradle
@@ -6,6 +6,9 @@ include ':kuick-repositories-elasticsearch'
 include ':kuick-repositories-squash'
 include ':kuick-redis'
 include ':kuick-cache-redis'
+include ':kuick-api'
+include ':kuick-api-rest'
+include ':kuick-api-rpc'
 
 //include ':kuick-postgres'
 include ':kuick-db'
